Clean up ALB config for Publisher and factor out domain name.
This introduces a new top-level var public_lb_domain_name which looks a
bit like a duplicate of govuk_app_domain_external as it's set to the
same thing ("test.govuk.digital") but they're for quite different
purposes and it's not clear yet whether the two things should be
logically identical, so leaving them separate for now for clarity and
principle of least surprise.

The name public_lb_domain_name probably isn't the best, but at least it
makes it obvious what it's for at the moment.

Also:
 - Enable evaluate_target_health.
 - Remove a redundant domain suffix from a Route53 resource (this was a
   common antipattern in govuk-aws so was probably copied from there).
 - Improve some descriptions.