Add deployment module for content-store
This will enable a deployment pipeline to use Terraform to create a
new task definition revision for content-store in the test environment.

A common.tf file holds the locals and variables required to create
new task definitions.

This could probably be refactored a bit to use tfvars files, rather
than nesting the apps under a test directory.

Co-authored-by: Stephen Ford <stephen.ford@digital.cabinet-office.gov.uk>