Add Concourse pipeline for deploying apps to ECS
This is a minimal implementation of the deployment pipeline proposed
in the [deployments design doc][] (not public) and [proposal 005][]
(public).

The single Concourse pipeline watches GitHub repos for changes to the main branches
and deploys to ECS in an AWS account using a Terraform app deployment module.

Significantly more work is required for this pipeline to meet the requirements
for deployment. We will need to add features for rollbacks, checking smoke
tests, permitting branch deploys, restricting deploy permissions and access,
and running deploy hooks such as database migrations. These features are feasible
but will require iterating on this implementation.

[deployments design doc]: https://docs.google.com/document/d/1qmNK7TqR0_aR5i_JRITchsutTAEZavmzGs63LI1XNbA/edit
[proposal 005]: https://github.com/alphagov/govuk-replatforming-discovery-2020/blob/b09454e2462e0c7cfcbc7c6a0ebf1760b693e23b/proposals/005-ecs-fargate-deployment.md

Story: https://trello.com/c/C2ubxCmM/175-minimal-deployment-solution

Co-authored-by: Stephen Ford <stephen.ford@digital.cabinet-office.gov.uk>