Run Terraform plan in Concourse for main branch
We'd like to run a Terraform plan for every branch on govuk-infrastructure
before merging the changes into the main branch.

This creates a Concourse pipeline called ci-govuk-infrastructure-main
that will be used to test that a Terraform plan does not raise errors
against the main branch.

This doesn't yet take into account the workspaces work, but we should
probably run a plan against the default workspace / cluster - to ensure
we can apply the main branch to the default workspace.