Update secrets baseline
I think people have probably got a bit lazy with running this by hand (I
certainly have), so there are a couple of places where it's drifted
out-of-date.

This updates it to match the current state of the repo.

I'll add a follow up to add pre-commit to suggest that people use the
pre-commit hook.