Add Statsd ECS Service
This recreates a Statsd ECS Service that we had during the discovery
phase of the project. We're going to get this Statsd instance
to send metrics to Graphite in EC2, so we don't need to recreate our
app metrics in Prometheus immediately.

Sadly App Mesh doesn't support UDP so we're still using TCP
to send metrics to Statsd.

This factors out the App Mesh and CloudMap resources into a
service-mesh-node module, and shares that module between statsd
and the app module. The intent is to make statsd not be treated
as an app, since it doesn't have a deployment requirement.