Enable apps to ship statsd metrics
This security group rule will permit Statsd to receive
metrics from any ECS Service created in the app module.

This also adds a security group which will be added to every
virtual service (ECS Service) in the service mesh (App Mesh).
This security group will be useful since we expect to have
multiple services that we wish most/all apps in ECS to send
requests to (statsd, prometheus) so this should reduce the
number of groups/rules we need to create.

Co-Authored-By: Bill Franklin <william.franklin@digital.cabinet-office.gov.uk>