Add fetch task definition Concourse task
This task will be useful when we want to pull the task definition
and other outputs from an app deployment, but do not want to apply
the deployment.

For example, when running smoke tests in parallel, we don't want
to attempt creation of a smoke test ECS Task Definition for every
test run - we just want to use the latest smoke test task definition
according to the Terraform deployment.