Run smokey tests in ECS after frontend deploys
This will trigger a smokey run in ECS after every deploy of the
frontend app.

Currently we run a subset of the test suite in the test environment
called 'replatforming', which are just the tests we expect to run
successfully.

In future, I expect we'll change this to 'app-frontend', which will
run all the smoke tests that would be likely to fail should frontend
introduce a bug or some other issue. Currently most frontend tests
fail.

We may also prefer to run smoke tests based on the domain, rather than
on an app. These should be expressed in smokey test tags, and updated
here.