Add task definition v2 and container definition modules
This introduces two modules for generating a task definition
JSON, moving the logic out of modules/app/task_definition.tf.

This enables us to register task definitions using the same
config without also needing to create an ECS Service or run
an Envoy proxy in the container.
This is useful for scheduled tasks, such as Smokey.