Automate creation and rotation of Publisher's bearer token for Publishing API
This uses the signon bearer token rotation lambda to create a bearer token
for Publisher to use in requests to Publishing API.

The lambda creates a bearer token, and will rotate the token (working with
signon and secrets manager to achieve this) every 30 days.

The motivation for this change is that manually creating bearer tokens in a
new environment is time consuming. This approach also lends itself to infrequent
bearer token rotation schedules, which is bad for our security posture.

Further work is required to fully automate bootstrapping signon.